Refactored upload
The upload() method tries to upload a file and will raise an exception
when it fails.  Instead, to upload or update an existing file, the
upload_or_update() method should be used.  It returns a tuple of the
returned JSON and an `uploaded` boolean describing whether the file was
uploaded; when False, the file was updated.